\name{paul-package}
\alias{paul-package}
\alias{paul}
\docType{package}
\title{
\packageTitle{paul}
}
\description{
\packageDescription{paul}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{paul}
\packageIndices{paul}
~~ An overview of how to use the package, ~~
~~ including the most important functions ~~
}
\author{
\packageAuthor{paul}

Maintainer: \packageMaintainer{paul}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per ~~
~~ line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
